Corrections made based on the comments in the PR.
